add geolocation and minecraft locale detection
