--- README.md
@@ -1,2 +1,0 @@
-# LanguageAPI
-LanguageAPI sources
