+++ pom.xml
@@ -1,0 +1,82 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <groupId>fr.tenebrae.PlayerLanguage</groupId>
+  <artifactId>LanguageAPI</artifactId>
+  <version>0.0.1-SNAPSHOT</version>
+  <name>PixelAPI</name>
+  <description>LanguageAPI</description>
+  <properties>
+    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+  </properties>
+  <build>
+    <finalName>LanguageAPI</finalName>
+    <sourceDirectory>src/main/java</sourceDirectory>
+    <resources>
+		<resource>
+			<filtering>true</filtering>
+			<directory>src/main/resources</directory>
+			<includes>
+				<include>plugin.yml</include>
+				<include>config.yml</include>
+				<include>README.md</include>
+				<include>GeoLiteCity.dat</include>
+			</includes>
+		</resource>
+	</resources>
+    <plugins>
+    <plugin>
+	  <artifactId>maven-assembly-plugin</artifactId>
+	  <configuration>
+	    <executable>C:\Program Files\Java\jdk1.7.0_79\bin\javac.exe</executable>
+	    <archive>
+	      <manifest>
+	        <mainClass>fully.qualified.MainClass</mainClass>
+	      </manifest>
+	    </archive>
+	    <descriptorRefs>
+	  "src/main/java/org/pixeledge/api/PixelEdge.java"    <descriptorRef>jar-with-dependencies</descriptorRef>
+	    </descriptorRefs>
+	  </configuration>
+	  <executions>
+	    <execution>
+	      <id>make-assembly</id> <!-- this is used for inheritance merges -->
+	      <phase>package</phase> <!-- bind to the packaging phase -->
+	      <goals>
+	        <goal>single</goal>
+	      </goals>
+	    </execution>
+	  </executions>
+	  </plugin>
+      <plugin>
+        <artifactId>maven-compiler-plugin</artifactId>
+        <version>3.3</version>
+        <configuration>
+          <fork>true</fork>
+          <executable>C:\Program Files\Java\jdk1.7.0_79\bin\javac.exe</executable>
+          <source>1.7</source>
+          <target>1.7</target>
+        </configuration>
+      </plugin>
+    </plugins>
+  </build>
+  <repositories>
+      <repository>
+        <id>spigot-repo</id>
+        <url>https://hub.spigotmc.org/nexus/content/repositories/snapshots/</url>
+   	 </repository>
+  </repositories>
+	<dependencies>
+
+	<dependency>
+	    <groupId>org.spigotmc</groupId>
+	    <artifactId>spigot</artifactId>
+	    <version>1.9.2-R0.1-SNAPSHOT</version>
+	</dependency>
+	<dependency>
+        <groupId>com.maxmind.geoip2</groupId>
+        <artifactId>geoip2</artifactId>
+        <version>2.7.0</version>
+    </dependency>
+		
+	</dependencies>
+</project>
