--- src/fr/tenebrae/PlayerLanguage/Languages.java
@@ -1,15 +1,0 @@
-package fr.tenebrae.PlayerLanguage;
-
-public enum Languages {
-	
-	FRENCH,
-	ENGLISH,
-	DEUTSCH,
-	SPANISH,
-	DUTCH,
-	PORTOUGUESE,
-	POLISH,
-	ARABISH,
-	ITALIAN,
-	CHINESE;
-}
