+++ src/main/java/fr/tenebrae/PlayerLanguage/Languages.java
@@ -1,0 +1,15 @@
+package fr.tenebrae.PlayerLanguage;
+
+public enum Languages {
+	
+	FRENCH,
+	ENGLISH,
+	DEUTSCH,
+	SPANISH,
+	DUTCH,
+	PORTOUGUESE,
+	POLISH,
+	ARABISH,
+	ITALIAN,
+	CHINESE;
+}
